requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
autobahn==23.6.2; python_version >= '3.9'
automat==22.10.0
certifi==2024.6.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
channels[daphne]==4.1.0; python_version >= '3.8'
channels-redis==4.2.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
constantly==23.10.4; python_version >= '3.8'
cryptography==42.0.7; python_version >= '3.7'
daphne==4.1.2
defusedxml==0.8.0rc2; python_version >= '3.6'
dj-database-url==2.2.0
django==5.0.6; python_version >= '3.10'
django-debug-toolbar==4.4.2; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
django-templated-mail==1.1.1
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djoser==2.2.2; python_version >= '3.8' and python_version < '4.0'
drf-nested-routers==0.94.1; python_version >= '3.8'
gunicorn==22.0.0; python_version >= '3.7'
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.0; python_version >= '3.7'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pyopenssl==24.1.0
python3-openid==3.2.0
redis==5.0.4; python_version >= '3.7'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
service-identity==24.1.0
setuptools==70.0.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.4.1; python_version >= '3.8'
social-auth-core==4.5.4; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
twisted[tls]==24.3.0; python_full_version >= '3.8.0'
txaio==23.1.1; python_version >= '3.7'
typing-extensions==4.12.1; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
zope-interface==6.4.post2; python_version >= '3.7'
